feat: support multiple quality media
