+++ fuocore/__init__.py
@@ -1,10 +1,11 @@
 from fuocore.models import ModelType  # noqa
 from fuocore.player import (
-    MpvPlayer,
     State as PlayerState,
     PlaybackMode,
     Playlist,
 )  # noqa
+# FIXME: remove this when no one import MpvPlayer from here
+from fuocore.mpvplayer import MpvPlayer  # noqa
 from fuocore.live_lyric import LiveLyric  # noqa
 from .library import Library  # noqa
